% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{create_exclusions_preview}
\alias{create_exclusions_preview}
\title{Starts the generation of an exclusions preview for the specified assessment template}
\usage{
create_exclusions_preview(
  assessmentTemplateArn = "string"
)
}
\arguments{
\item{assessmentTemplateArn}{[required] The ARN that specifies the assessment template for which you want to create an exclusions preview.}
}
\description{
Starts the generation of an exclusions preview for the specified assessment template. The exclusions preview lists the potential exclusions (ExclusionPreview) that Inspector can detect before it runs the assessment.
}
\examples{

}
