% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{delete_assessment_template}
\alias{delete_assessment_template}
\title{Deletes the assessment template that is specified by the ARN of the assessment template}
\usage{
delete_assessment_template(
  assessmentTemplateArn = "string"
)
}
\arguments{
\item{assessmentTemplateArn}{[required] The ARN that specifies the assessment template that you want to delete.}
}
\description{
Deletes the assessment template that is specified by the ARN of the assessment template.
}
\examples{
# Deletes the assessment template that is specified by the ARN of the
# assessment template.
delete_assessment_template(
  assessmentTemplateArn = "arn:aws:inspector:us-west-2:123456789012:target/0-nvgVhaxX/template/0-it5r2S4T"
)

}
