% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{describe_assessment_targets}
\alias{describe_assessment_targets}
\title{Describes the assessment targets that are specified by the ARNs of the assessment targets}
\usage{
describe_assessment_targets(assessmentTargetArns)
}
\arguments{
\item{assessmentTargetArns}{The ARNs that specifies the assessment targets that you want to describe.}
}
\description{
Describes the assessment targets that are specified by the ARNs of the assessment targets.
}
\examples{
# Describes the assessment targets that are specified by the ARNs of the
# assessment targets.
describe_assessment_targets(
  assessmentTargetArns = list(
    "arn:aws:inspector:us-west-2:123456789012:target/0-0kFIPusq"
  )
)

}
