% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{describe_assessment_templates}
\alias{describe_assessment_templates}
\title{Describes the assessment templates that are specified by the ARNs of the assessment templates}
\usage{
describe_assessment_templates(assessmentTemplateArns)
}
\arguments{
\item{assessmentTemplateArns}{[required]}
}
\description{
Describes the assessment templates that are specified by the ARNs of the assessment templates.
}
\section{Accepted Parameters}{
\preformatted{describe_assessment_templates(
  assessmentTemplateArns = list(
    "string"
  )
)
}
}

\examples{
# Describes the assessment templates that are specified by the ARNs of the
# assessment templates.
\donttest{describe_assessment_templates(
  assessmentTemplateArns = list(
    "arn:aws:inspector:us-west-2:123456789012:target/0-0kFIPusq/template/0-4r1V2mAw"
  )
)}

}
