% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{describe_cross_account_access_role}
\alias{describe_cross_account_access_role}
\title{Describes the IAM role that enables Amazon Inspector to access your AWS account}
\usage{
describe_cross_account_access_role()
}
\description{
Describes the IAM role that enables Amazon Inspector to access your AWS account.
}
\examples{
# Describes the IAM role that enables Amazon Inspector to access your AWS
# account.
describe_cross_account_access_role()

}
