% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{describe_resource_groups}
\alias{describe_resource_groups}
\title{Describes the resource groups that are specified by the ARNs of the resource groups}
\usage{
describe_resource_groups(resourceGroupArns)
}
\arguments{
\item{resourceGroupArns}{[required] The ARN that specifies the resource group that you want to describe.}
}
\description{
Describes the resource groups that are specified by the ARNs of the resource groups.
}
\section{Accepted Parameters}{
\preformatted{describe_resource_groups(
  resourceGroupArns = list(
    "string"
  )
)
}
}

\examples{
# Describes the resource groups that are specified by the ARNs of the
# resource groups.
\donttest{describe_resource_groups(
  resourceGroupArns = list(
    "arn:aws:inspector:us-west-2:123456789012:resourcegroup/0-PyGXopAI"
  )
)}

}
