% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{describe_rules_packages}
\alias{describe_rules_packages}
\title{Describes the rules packages that are specified by the ARNs of the rules packages}
\usage{
describe_rules_packages(
  rulesPackageArns = list(
    "string"
  ),
  locale = "EN_US"
)
}
\arguments{
\item{rulesPackageArns}{[required] The ARN that specifies the rules package that you want to describe.}

\item{locale}{The locale that you want to translate a rules package description into.}
}
\description{
Describes the rules packages that are specified by the ARNs of the rules packages.
}
\examples{
# Describes the rules packages that are specified by the ARNs of the rules
# packages.
describe_rules_packages(
  rulesPackageArns = list(
    "arn:aws:inspector:us-west-2:758058086616:rulespackage/0-JJOtZiqQ"
  )
)

}
