% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{get_exclusions_preview}
\alias{get_exclusions_preview}
\title{Retrieves the exclusions preview (a list of ExclusionPreview objects) specified by the preview token}
\usage{
get_exclusions_preview(assessmentTemplateArn, previewToken, nextToken = NULL,
  maxResults = NULL, locale = NULL)
}
\arguments{
\item{assessmentTemplateArn}{The ARN that specifies the assessment template for which the exclusions preview was requested.}

\item{previewToken}{The unique identifier associated of the exclusions preview.}

\item{nextToken}{You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the GetExclusionsPreviewRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.}

\item{maxResults}{You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.}

\item{locale}{The locale into which you want to translate the exclusion's title, description, and recommendation.}
}
\description{
Retrieves the exclusions preview (a list of ExclusionPreview objects) specified by the preview token. You can obtain the preview token by running the CreateExclusionsPreview API.
}
\examples{

}
