% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags associated with an assessment template}
\usage{
list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{The ARN that specifies the assessment template whose tags you want to list.}
}
\description{
Lists all tags associated with an assessment template.
}
\examples{
# Lists all tags associated with an assessment template.
list_tags_for_resource(
  resourceArn = "arn:aws:inspector:us-west-2:123456789012:target/0-0kFIPusq/template/0-gcwFliYu"
)

}
