% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_package.R
\docType{package}
\name{paws.inspector-package}
\alias{paws.inspector}
\alias{paws.inspector-package}
\title{paws.inspector: Amazon Inspector}
\description{
Amazon Inspector enables you to analyze the
    behavior of your AWS resources and to identify potential security
    issues. For more information, see [Amazon Inspector User
    Guide](http://docs.aws.amazon.com/inspector/latest/userguide/inspector_introduction.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
