% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{register_cross_account_access_role}
\alias{register_cross_account_access_role}
\title{Registers the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments}
\usage{
register_cross_account_access_role(roleArn)
}
\arguments{
\item{roleArn}{The ARN of the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments.}
}
\description{
Registers the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments.
}
\examples{
# Registers the IAM role that Amazon Inspector uses to list your EC2
# instances at the start of the assessment run or when you call the
# PreviewAgents action.
register_cross_account_access_role(
  roleArn = "arn:aws:iam::123456789012:role/inspector"
)

}
