% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{stop_assessment_run}
\alias{stop_assessment_run}
\title{Stops the assessment run that is specified by the ARN of the assessment run}
\usage{
stop_assessment_run(assessmentRunArn, stopAction = NULL)
}
\arguments{
\item{assessmentRunArn}{The ARN of the assessment run that you want to stop.}

\item{stopAction}{An input option that can be set to either START\_EVALUATION or SKIP\_EVALUATION. START\_EVALUATION (the default value), stops the AWS agent from collecting data and begins the results evaluation and the findings generation process. SKIP\_EVALUATION cancels the assessment run immediately, after which no findings are generated.}
}
\description{
Stops the assessment run that is specified by the ARN of the assessment run.
}
\examples{
# Stops the assessment run that is specified by the ARN of the assessment
# run.
stop_assessment_run(
  assessmentRunArn = "arn:aws:inspector:us-west-2:123456789012:target/0-nvgVhaxX/template/0-it5r2S4T/run/0-11LMTAVe"
)

}
