% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{update_assessment_target}
\alias{update_assessment_target}
\title{Updates the assessment target that is specified by the ARN of the assessment target}
\usage{
update_assessment_target(assessmentTargetArn, assessmentTargetName,
  resourceGroupArn = NULL)
}
\arguments{
\item{assessmentTargetArn}{The ARN of the assessment target that you want to update.}

\item{assessmentTargetName}{The name of the assessment target that you want to update.}

\item{resourceGroupArn}{The ARN of the resource group that is used to specify the new resource group to associate with the assessment target.}
}
\description{
Updates the assessment target that is specified by the ARN of the assessment target.
}
\details{
If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.
}
\examples{
# Updates the assessment target that is specified by the ARN of the
# assessment target.
update_assessment_target(
  assessmentTargetArn = "arn:aws:inspector:us-west-2:123456789012:target/0-nvgVhaxX",
  assessmentTargetName = "Example",
  resourceGroupArn = "arn:aws:inspector:us-west-2:123456789012:resourcegroup/0-yNbgL5Pt"
)

}
