% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.inspector_operations.R
\name{subscribe_to_event}
\alias{subscribe_to_event}
\title{Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic}
\usage{
subscribe_to_event(resourceArn, event, topicArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the assessment template that is used during the event for which you want to receive SNS notifications.}

\item{event}{[required] The event for which you want to receive SNS notifications.}

\item{topicArn}{[required] The ARN of the SNS topic to which the SNS notifications are sent.}
}
\description{
Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.
}
\section{Accepted Parameters}{
\preformatted{subscribe_to_event(
  resourceArn = "string",
  event = "ASSESSMENT_RUN_STARTED"|"ASSESSMENT_RUN_COMPLETED"|"ASSESSMENT_RUN_STATE_CHANGED"|"FINDING_REPORTED"|"OTHER",
  topicArn = "string"
)
}
}

\examples{
# Enables the process of sending Amazon Simple Notification Service (SNS)
# notifications about a specified event to a specified SNS topic.
\donttest{subscribe_to_event(
  event = "ASSESSMENT_RUN_COMPLETED",
  resourceArn = "arn:aws:inspector:us-west-2:123456789012:target/0-nvgVhaxX/template/0-7sbz2Kz0",
  topicArn = "arn:aws:sns:us-west-2:123456789012:exampletopic"
)}

}
